# api/swagger-api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/session": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Создание или обновление сессии",
        "description": "Если кука отсутствует или сессия истекла — создает новую (201). Если кука есть и сессия активна — обновляет её TTL в Redis (200).",
        "operationId": "session",
        "parameters": [
          {
            "name": "X-Session-Id",
            "in": "cookie",
            "description": "ID сессии из Cookie",
            "required": false,
            "schema": {
              "type": "string"
            },
            "example": "3f8a2c1d9e4b7f0a5c6d2e8b1a3f9c7d"
          }
        ],
        "responses": {
          "201": {
            "description": "Сессия создана (первый визит)",
            "headers": {
              "Set-Cookie": {
                "description": "X-Session-Id={sid}; HttpOnly; Path=/; Max-Age={APP_USER_SESSION_TTL}",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "200": {
            "description": "Сессия обновлена (повторный визит)",
            "headers": {
              "Set-Cookie": {
                "description": "X-Session-Id={sid}; HttpOnly; Path=/; Max-Age={APP_USER_SESSION_TTL}",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}